+++ src/main/java/com/kalsym/locationservice/repository/ProductRepository.java
@@ -23,13 +23,13 @@ public interface ProductRepository extends JpaRepository<ProductMain,String> {
         + "AND pwd.storeDetails.regionCountryId = :regionCountryId "
         + "AND pwd.status IN :status "
         + "AND pwd.storeDetails.regionCityDetails.name LIKE CONCAT('%', :cityName ,'%') "
-        + "AND pwd.storeDetails.city LIKE CONCAT('%', :cityId ,'%')"
+        + "AND pwd.storeDetails.city IN :cityId"
     )
     Page<ProductMain> getProductByParentCategoryIdAndLocation(
             @Param("status") List<String> status,
             @Param("regionCountryId") String regionCountryId,
             @Param("parentCategoryId") String parentCategoryId,
-            @Param("cityId") String cityId,
+            @Param("cityId") List<String> cityId,
             @Param("cityName") String cityName,
             @Param("name") String name,
             Pageable pageable