+++ src/main/java/com/kalsym/locationservice/repository/StoreConfigRepository.java
@@ -5,6 +5,8 @@ import com.kalsym.locationservice.model.Config.StoreConfig;
 
 import org.springframework.data.domain.Pageable;
 
+import java.util.List;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
@@ -22,25 +24,17 @@ public interface StoreConfigRepository extends JpaRepository<StoreConfig,Integer
         + "INNER JOIN Category sc on sc.storeId = sdc.storeId "
         + "WHERE s.regionCountryId LIKE CONCAT('%', :regionCountryId ,'%') "
         + "AND sc.parentCategoryId LIKE CONCAT('%', :parentCategoryId ,'%') "
-        + "AND s.city LIKE CONCAT('%', :cityId ,'%') "
+        + "OR s.city IN :cityId "
         + "AND s.regionCityDetails.name LIKE CONCAT('%', :cityName ,'%') "
         + "GROUP BY sdc.id"
     )
     Page<StoreConfig> getQueryStoreConfigRaw(
-        @Param("cityId") String cityId,
+        @Param("cityId") List<String> cityId,
         @Param("cityName") String cityName,
         @Param("regionCountryId") String regionCountryId,
         @Param("parentCategoryId") String parentCategoryId,
         Pageable pageable
     );
 }
-// Page<StoreConfig> result = storeConfigRepository.getQueryStoreConfigRaw
-// (cityId,
-// cityName,
-// stateId,
-// regionCountryId,
-// postcode,
-// parentCategoryId,
-// storeName,
-// pageable);
+